+++ src/interpreter.rs
@@ -1,4 +1,3 @@
-//use core::panicking::panic;
 use std::panic::panic_any;
 // use std::cell::RefCell;
 use std::collections::HashMap;
@@ -10,7 +9,11 @@ use std::ops::Deref;
 use std::process::id;
 use std::string::String;
 
-mod heap;
+#[derive(Debug, Clone)]
+pub struct Environment {
+    store: HashMap<String, u64>,
+    pool: Vec<Literal>
+}
 
 #[derive(Debug, Clone)]
 struct TopLevelMap {
@@ -41,15 +44,20 @@ struct Assignment_i {
 }
 
 #[derive(Debug, Clone)]
+struct Mark {
+    mem: u64
+}
+
+#[derive(Debug, Clone)]
 pub enum Instructions {
     Reset,
+    Mark(Mark),
     Assignment(Assignment),
     UnOp(UnOp),
     BinOp(BinOp),
     Pop,
     App,
     Branch,
-    Env,
     Assignment_i(Assignment_i)
 }
 
@@ -60,8 +68,10 @@ pub enum AgendaInstrs {
     Block(Block),
     SequenceStmt(SequenceStmt),
     Literal(Literal),
-    Expr(Expr)
+    Expr(Expr),
+    PrimitiveOperation(PrimitiveOperation)
 }
+
 
 /***************************************************************************************************
 * Run the code
@@ -147,18 +157,24 @@ pub fn run(ast: &mut Vec<parser::ast::Stmt>) {
     // A is our stack of instructions as in the Source EC evaluator
     let mut A: Vec<AgendaInstrs> = Vec::new();
     // S is the stash of evaluated values
-    let mut S: Vec<Literal> = Vec::new(); // TODO: Environments
+    let mut S: Vec<Literal> = Vec::new();
+    // E is the environment
+    let mut E = Environment {
+        store: HashMap::new(),
+        pool: vec![]
+    };
     // We start with main
-    A.push(AgendaInstrs::Stmt(functions.list[main_idx].clone())); // TODO: Push only main's block onto agenda
+    A.push(AgendaInstrs::Stmt(functions.list[main_idx].clone()));
     while !A.is_empty() {
         let curr: AgendaInstrs = A.pop().expect("Literally impossible but ok");
-        curr.evaluate(&mut A, &mut S); // Borrowing w/ Mutable Reference
+        curr.evaluate(&mut A, &mut S, &mut E); // Borrowing w/ Mutable Reference
     }
 }
+
 /***************************************************************************************************
 * Operators
 ***************************************************************************************************/
-pub fn binop_microcode_num(x: i64, y: i64, sym: BinaryOperator) -> crate::interpreter::Literal {
+pub fn binop_microcode_num(x: i64, y: i64, sym: BinaryOperator) -> Literal {
     let output =  match sym {
         BinaryOperator::Plus => x + y,
         BinaryOperator::Minus => x - y,
@@ -169,7 +185,7 @@ pub fn binop_microcode_num(x: i64, y: i64, sym: BinaryOperator) -> crate::interp
     return IntLiteral(output);
 }
 
-pub fn binop_microcode_num_bool(x: i64, y: i64, sym: BinaryOperator) -> crate::interpreter::Literal {
+pub fn binop_microcode_num_bool(x: i64, y: i64, sym: BinaryOperator) -> Literal {
     let output = match sym {
         BinaryOperator::Equal => x == y,
         BinaryOperator::NotEqual => x != y,
@@ -182,7 +198,7 @@ pub fn binop_microcode_num_bool(x: i64, y: i64, sym: BinaryOperator) -> crate::i
     return BoolLiteral(output);
 }
 
-pub fn binop_microcode_bool(x: bool, y: bool, sym: BinaryOperator) -> crate::interpreter::Literal {
+pub fn binop_microcode_bool(x: bool, y: bool, sym: BinaryOperator) -> Literal {
     let output =  match sym {
         BinaryOperator::And => x && y,
         BinaryOperator::Or => x || y,
@@ -191,30 +207,29 @@ pub fn binop_microcode_bool(x: bool, y: bool, sym: BinaryOperator) -> crate::int
     return BoolLiteral(output);
 }
 
-// Need to check if both x and y have matching types - FIXED
-pub fn apply_binop(x: Literal, y: Literal, sym: BinaryOperator) -> Literal {
+pub fn apply_binop(x: Option<Literal>, y: Option<Literal>, sym: BinaryOperator) -> Literal {
     let output = match x { // Check type of x
-      Literal::IntLiteral(x_num) => match y {
-          Literal::IntLiteral(y_num) => match sym { // Ensure x & y are both integer types
-              BinaryOperator::Plus|BinaryOperator::Minus|BinaryOperator::Divide|BinaryOperator::Times => binop_microcode_num(x_num, y_num, sym),
-              BinaryOperator::Equal|BinaryOperator::NotEqual|BinaryOperator::Greater|BinaryOperator::GreaterOrEqual|BinaryOperator::Less|BinaryOperator::LessOrEqual => binop_microcode_num_bool(x_num, y_num, sym),
-              _ => panic!("fn. apply_binop operator unsupported for types x(int) and y(int)!")
-          }
-          _ => panic!("fn. apply_binop x(int) and y have different types!")
-      }
-      Literal::BoolLiteral(x_bool) => match y {
-          Literal::BoolLiteral(y_bool) => match sym { // Ensure x any y are both bool types
-              BinaryOperator::And|BinaryOperator::Or => binop_microcode_bool(x_bool, y_bool, sym),
-              _ => panic!("fn. apply_binop operator unsupported for types x(bool) and y(bool)!")
-          }
-          _ => panic!("fn. apply_binop x(bool) and y have different types!")
-      }
-      _=> panic!("fn. apply_binop primitive operations unsupported for this type!")
+        Some(Literal::IntLiteral(x_num)) => match y {
+            Some(Literal::IntLiteral(y_num)) => match sym { // Ensure x & y are both integer types
+                BinaryOperator::Plus|BinaryOperator::Minus|BinaryOperator::Divide|BinaryOperator::Times => binop_microcode_num(x_num, y_num, sym),
+                BinaryOperator::Equal|BinaryOperator::NotEqual|BinaryOperator::Greater|BinaryOperator::GreaterOrEqual|BinaryOperator::Less|BinaryOperator::LessOrEqual => binop_microcode_num_bool(x_num, y_num, sym),
+                _ => panic!("fn. apply_binop operator unsupported for types x(int) and y(int)!")
+            }
+            _ => panic!("fn. apply_binop x(int) and y have different types!")
+        }
+        Some(Literal::BoolLiteral(x_bool)) => match y {
+            Some(Literal::BoolLiteral(y_bool)) => match sym { // Ensure x any y are both bool types
+                BinaryOperator::And|BinaryOperator::Or => binop_microcode_bool(x_bool, y_bool, sym),
+                _ => panic!("fn. apply_binop operator unsupported for types x(bool) and y(bool)!")
+            }
+            _ => panic!("fn. apply_binop x(bool) and y have different types!")
+        }
+        _=> panic!("fn. apply_binop primitive operations unsupported for this type!")
     };
     return output;
 }
 
-pub fn unop_microcode_bool(x: bool, sym: UnaryOperator) -> crate::interpreter::Literal {
+pub fn unop_microcode_bool(x: bool, sym: UnaryOperator) -> Literal {
     let output = match sym {
         UnaryOperator::Not => !x,
         _=> panic!("fn. unop_microcode_bool unsupported operator!")
@@ -222,7 +237,7 @@ pub fn unop_microcode_bool(x: bool, sym: UnaryOperator) -> crate::interpreter::L
     return BoolLiteral(output);
 }
 
-pub fn unop_microcode_num(x: i64, sym: UnaryOperator) -> crate::interpreter::Literal {
+pub fn unop_microcode_num(x: i64, sym: UnaryOperator) -> Literal {
     let output = match sym {
         UnaryOperator::UnaryMinus => -x,
         _=> panic!("fn. unop_microcode_num unsupported operator!")
@@ -230,23 +245,13 @@ pub fn unop_microcode_num(x: i64, sym: UnaryOperator) -> crate::interpreter::Lit
     return IntLiteral(output);
 }
 
-pub fn apply_unop(x: Literal, sym: UnaryOperator) -> Literal {
+pub fn apply_unop(x: Option<Literal>, sym: UnaryOperator) -> Literal {
     let output = match x {
-        Literal::IntLiteral(value) => unop_microcode_num(value, sym),
-        Literal::BoolLiteral(value) => unop_microcode_bool(value, sym),
+        Some(Literal::IntLiteral(value)) => unop_microcode_num(value, sym),
+        Some(Literal::BoolLiteral(value)) => unop_microcode_bool(value, sym),
         _ => panic!("fn. apply_binop unsupported type for x!")
     };
     return output;
-}
-
-/***************************************************************************************************
-* Get the identifier name
-***************************************************************************************************/
-fn get_name(expr: &Expr) -> String {
-    match expr {
-        Expr::IdentifierExpr(name, sourcelocation) => name.clone(),
-        _ => panic!("fn. get_name could not find identifier to get name from")
-    }
 }
 
 /***************************************************************************************************
@@ -289,6 +294,15 @@ fn scan_out_block_declarations(block: &Block) -> Vec<String> {
 }
 
 /***************************************************************************************************
+* Get the identifier name
+***************************************************************************************************/
+fn get_name(expr: &Expr) -> String {
+    match expr {
+        Expr::IdentifierExpr(name, sourcelocation) => name.clone(),
+        _ => panic!("fn. get_name could not find identifier to get name from")
+    }
+}
+/***************************************************************************************************
 * Stack
 ***************************************************************************************************/
 trait Stack {
@@ -303,111 +317,208 @@ impl Stack for Vec<Literal> {
         }
     }
 }
-
 /***************************************************************************************************
 * Evaluation
 ***************************************************************************************************/
 pub trait Evaluate {
-    fn evaluate(&self, instr_stack: &mut Vec<AgendaInstrs>, stash: &mut Vec<Literal>);
+    fn evaluate(&self, instr_stack: &mut Vec<AgendaInstrs>, stash: &mut Vec<Literal>, env: &mut Environment);
 }
 
 impl Evaluate for AgendaInstrs {
-    fn evaluate(&self, instr_stack: &mut Vec<AgendaInstrs>, stash: &mut Vec<Literal>) {
+    fn evaluate(&self, instr_stack: &mut Vec<AgendaInstrs>, stash: &mut Vec<Literal>, env: &mut Environment) {
         match self {
-            AgendaInstrs::Stmt(stmt) => stmt.evaluate(instr_stack, stash),
-            AgendaInstrs::SequenceStmt(stmts) => stmts.evaluate(instr_stack, stash),
-            AgendaInstrs::Block(blk) => blk.evaluate(instr_stack, stash),
-            AgendaInstrs::Literal(lit) => lit.evaluate(instr_stack, stash),
-            AgendaInstrs::Expr(expr) => expr.evaluate(instr_stack, stash),
-            AgendaInstrs::Instructions(instr) => { match instr {
-                Instructions::Reset => {},
-                Instructions::Assignment(assn) => {
-                    let a = Assignment_i {
-                        sym: assn.clone().sym
-                    };
-                    instr_stack.push(AgendaInstrs::Instructions(Instructions::Assignment_i(a)));
-                    instr_stack.push(AgendaInstrs::Expr(assn.clone().expr));
-                },
-                Instructions::UnOp(unop) => {},
-                Instructions::BinOp(binop) => {},
-                Instructions::Pop => {},
-                Instructions::App => {},
-                Instructions::Branch => {},
-                Instructions::Env => {},
-                Instructions::Assignment_i(assn) => {
-                    // Peek top of stash
-                    let v = match stash.peek() {
-                        Some(value) => *(value.clone()),
-                        None => panic!("Why is nothing in the stash??")
-                    };
-                    // Assign the value to the name in the environment TODO
+            AgendaInstrs::Stmt(stmt) => stmt.evaluate(instr_stack, stash, env),
+            // AgendaInstrs::SequenceStmt(stmts) => stmts.evaluate(instr_stack, stash),
+            AgendaInstrs::PrimitiveOperation(prim_op) => prim_op.evaluate(instr_stack, stash, env),
+            AgendaInstrs::Block(blk) => blk.evaluate(instr_stack, stash, env),
+            AgendaInstrs::Literal(lit) => lit.evaluate(instr_stack, stash, env),
+            AgendaInstrs::Expr(expr) => expr.evaluate(instr_stack, stash, env),
+            AgendaInstrs::Instructions(instr) => {
+                match instr {
+                    Instructions::Reset => {
+                        match instr_stack.pop() {
+                            Some(AgendaInstrs::Instructions(Instructions::Mark(m))) => {
+                                // TODO: pop_environment and mutate the current environment
+                            }, // Mark found?
+                            _ => instr_stack.push(AgendaInstrs::Instructions(Instructions::Reset)) // Continue loop by pushing reset back onto agenda
+                        }
+                    },
+                    Instructions::Assignment(assn) => {
+                        let a = Assignment_i {
+                            sym: assn.clone().sym
+                        };
+                        instr_stack.push(AgendaInstrs::Instructions(Instructions::Assignment_i(a)));
+                        instr_stack.push(AgendaInstrs::Expr(assn.clone().expr));
+                    },
+                    Instructions::UnOp(unop) => {
+                        let operand = stash.pop();
+                        let operator = unop.sym;
+                        let value = apply_unop(operand, operator);
+                        stash.push(value);
+                    },
+                    Instructions::BinOp(binop) => {
+                        let lhs_operand = stash.pop();
+                        let rhs_operand = stash.pop();
+                        let operator = binop.sym;
+                        let value = apply_binop(lhs_operand, rhs_operand, operator);
+                        println!("{:#?}", value.clone());
+                        stash.push(value);
+                    },
+                    Instructions::Pop => {
+                        stash.pop();
+                        println!("Pop!!")
+                    },
+                    Instructions::App => {},
+                    Instructions::Branch => {},
+                    Instructions::Mark(m) => {
+                        // TODO: pop_environment and mutate the current environment
+                    },
+                    Instructions::Assignment_i(assn) => {
+                        let v = match stash.peek() {
+                            Some(value) => value.clone(),
+                            None => panic!("Why is nothing in the stash??")
+                        };
+                        // Assign the value to the name in the environment
+                        let nam = assn.clone().sym;
+                        match env.store.get(&nam) { // Get index of where it is stored in the pool
+                            Some(val) => {
+                                env.pool.insert(*val as usize, v); // Assign value to variable
+                            }
+                            None => {
+                                panic!("Variable not initialised in the environment!");
+                            }
+                        }
+                    }
+                    _ => println!("Nothing is here")
                 }
-            }}
+            }
+            _ => println!("Nothing is here")
         }
     }
 }
 
+
 impl Evaluate for Stmt {
-    fn evaluate(&self, instr_stack: &mut Vec<AgendaInstrs>, stash: &mut Vec<Literal>) {
+    fn evaluate(&self, instr_stack: &mut Vec<AgendaInstrs>, stash: &mut Vec<Literal>, env: &mut Environment) {
         match self {
-            Stmt::LetStmt { name, is_mutable, annotation, value, position } => match value {
-               Some(expr) => {
-                  let name = get_name(name);
-                   // TODO: Put name in the environment
-                   instr_stack.push(AgendaInstrs::Literal(Literal::UnitLiteral));
-                   instr_stack.push(AgendaInstrs::Instructions(Instructions::Pop));
-                   let a = Assignment {
-                       sym: name.clone(),
-                       expr: expr.clone()
-                   };
-                   instr_stack.push(AgendaInstrs::Instructions(Instructions::Assignment(a)));
-               },
+            Stmt::LetStmt {name, is_mutable, annotation, value, position} => match value {
+                Some(expr) => {
+                    let name = get_name(name);
+                    // TODO: Put name in the environment
+                    instr_stack.push(AgendaInstrs::Literal(Literal::UnitLiteral));
+                    instr_stack.push(AgendaInstrs::Instructions(Instructions::Pop));
+                    let a = Assignment {
+                        sym: name.clone(),
+                        expr: expr.clone()
+                    };
+                    instr_stack.push(AgendaInstrs::Instructions(Instructions::Assignment(a)));
+                },
                 None => panic!("Unbounded declaration is currently unsupported!")
             },
             Stmt::FuncDeclaration { name, lifetime_parameters, parameters, return_type, body, position } => {
-                let no_params = parameters.len();
-                parameters
-                    .iter()
-                    .map(|(expr, _) | get_name(expr))
-                    .collect::<Vec<String>>()
-                    .into_iter()
-                    .for_each(|name| {
-                        // Add each variable to the env
-                    });
-                // TODO Fix
+                // let no_params = parameters.len();
+                // parameters
+                //     .iter()
+                //     .map(|(expr, _) | get_name(expr))
+                //     .collect::<Vec<String>>()
+                //     .into_iter()
+                //     .for_each(|name| {
+                //         // Add each variable to the env
+                //     });
+                // TODO: Fix
+                instr_stack.push(AgendaInstrs::Block(body.clone()));
             },
             Stmt::ExprStmt(expr) => match expr {
                 Expr::ReturnExpr(expr, loc) => {
+                    println!("In the return stmt");
                     instr_stack.push(AgendaInstrs::Instructions(Instructions::Reset));
-                    instr_stack.push(AgendaInstrs::Expr(**expr));
+                    let expr_clone = expr.clone();
+                    instr_stack.push(AgendaInstrs::Expr(*expr_clone));
                 },
                 _ => {
                     instr_stack.push(AgendaInstrs::Literal(Literal::UnitLiteral));
                     instr_stack.push(AgendaInstrs::Instructions(Instructions::Pop));
-                    instr_stack.push(AgendaInstrs::Expr(*(expr.clone())));
+                    instr_stack.push(AgendaInstrs::Expr((expr.clone())));
                 }
             },
-            _ => panic!("Unsupported statement")
+            _ => {
+                println!("Not a function");
+            }
         }
     }
 }
 
+impl Evaluate for Block {
+    fn evaluate(&self, instr_stack: &mut Vec<AgendaInstrs>, stash: &mut Vec<Literal>, env: &mut Environment) {
+        let old_env = Mark { // TODO: store the previous env on the heap with mark pointing to that mem location
+            mem: 0
+        };
+        instr_stack.push(AgendaInstrs::Instructions(Instructions::Mark(old_env))); // Drop mark on agenda
+
+        let mut locals = scan_out_block_declarations(self);
+        let mut curr_local: Option<String> = locals.pop();
+        while curr_local.is_some() {
+            // Overwrite current env values if there are similarly named variables
+            let curr: String = curr_local.expect("No locals");
+            match env.store.get(&curr) { // Get index of where it is stored in the pool
+                Some(val) => { // If the map points to some memory in the pool already
+                    env.pool.insert(*val as usize, Literal::UnitLiteral); // Unassigned for now
+                }
+                None => { // Else assign a new spot in the pool and then map the identifier to that
+                    env.pool.push(Literal::UnitLiteral);
+                    env.store.insert(curr, (env.pool.len() - 1) as u64);
+                }
+            }
+            curr_local = locals.pop();
+        }
+
+        let mut statements_clone = self.statements.clone();
+        let mut curr_stmt: Option<SequenceStmt> = statements_clone.pop();
+        while curr_stmt.is_some() {
+            let curr: SequenceStmt = curr_stmt.expect("No block statements?"); // current statement
+            match curr.clone() {
+                SequenceStmt::Stmt(s) => {
+                    instr_stack.push(AgendaInstrs::Stmt(s));
+                    instr_stack.push(AgendaInstrs::Instructions(Instructions::Pop));
+                }
+                SequenceStmt::Block(b) => instr_stack.push(AgendaInstrs::Block(b))
+            }
+            curr_stmt = statements_clone.pop();
+        }
+        instr_stack.pop(); // Remove the last pop that was put
+    }
+}
 
 impl Evaluate for Expr {
-    fn evaluate(&self, instr_stack: &mut Vec<AgendaInstrs>, stash: &mut Vec<Literal>) {
+    fn evaluate(&self, instr_stack: &mut Vec<AgendaInstrs>, stash: &mut Vec<Literal>, env: &mut Environment) {
         match self {
             Expr::IdentifierExpr(name, source_location) => {
-                // Find the identifier in the environment
-                // Push the identifier onto the stash. (Placeholder for now)
-                stash.push(Literal::IntLiteral(0)); // TODO: Complete this.
+                match env.store.get(name) { // Find identifier in the environment
+                    Some(idx) => {
+                        let value = match env.pool.get((*idx) as usize) { // Find identifier in pool
+                            Some(lit) => {
+                                lit.clone()
+                            }
+                            None => {
+                                panic!("evaluate identifier expr: Identifier is not in pool!");
+                            }
+                        };
+                        stash.push(value);
+                    }
+                    None => {
+                        panic!("evaluate identifier expr: Identifier not found in environment!");
+                    }
+                }
             }
             Expr::LiteralExpr(literal_value, source_location) => {
-                literal_value.evaluate(instr_stack, stash);
+                literal_value.evaluate(instr_stack, stash, env);
             }
-            Expr::BlockExpr(block, source_location) => {
-                block.evaluate(instr_stack, stash);
+            Expr::BlockExpr(block, source_location) => { // Block expr need to have a return stmt
+                block.evaluate(instr_stack, stash, env);
             }
             Expr::PrimitiveOperationExpr(primitive_op, source_location) => {
+                let prim_op = *primitive_op.clone();
+                instr_stack.push(AgendaInstrs::PrimitiveOperation(prim_op));
             }
             Expr::AssignmentExpr { assignee, value, position } => {}
             Expr::ApplicationExpr { is_primitive, callee, arguments, position } => {}
@@ -416,101 +527,42 @@ impl Evaluate for Expr {
     }
 }
 
-impl Evaluate for SequenceStmt {
-    fn evaluate(&self, instr_stack: &mut Vec<AgendaInstrs>, stash: &mut Vec<Literal>) {
-        match self {
-            SequenceStmt::Stmt(stmt) => stmt.evaluate(instr_stack, stash),
-            SequenceStmt::Block(block) => block.evaluate(instr_stack, stash),
-        }
-    }
-}
-
-impl Evaluate for Block {
-    fn evaluate(&self, instr_stack: &mut Vec<AgendaInstrs>, stash: &mut Vec<Literal>) {
-        let locals = scan_out_block_declarations(self);
-        // TODO: Locals into the environment, as unassigned
-        // TODO: Push environment onto agenda
-
-        // TODO: Push each statement onto the agenda (reverse order),similar to handle_sequence in ec-evaluator
-        let instrs = vec![];
-        let first = true;
-        if (&self.statements.len() == 0) {
-            instr_stack.push()
-        }
-        for stmt in &self.statements {
-
-        }
-    }
-}
-
 impl Evaluate for PrimitiveOperation{
-    fn evaluate(&self, instr_stack: &mut Vec<AgendaInstrs>, stash: &mut Vec<Literal>) {
+    fn evaluate(&self, instr_stack: &mut Vec<AgendaInstrs>, stash: &mut Vec<Literal>, env: &mut Environment) {
         match self {
             PrimitiveOperation::UnaryOperation { operator, operand } => {
-                operand.evaluate(instr_stack, stash);
-                operator.evaluate(instr_stack, stash);
+                let un_operator = UnOp {
+                    sym: operator.clone()
+                };
+                let un_operand = operand.clone();
+                instr_stack.push(AgendaInstrs::Instructions(Instructions::UnOp(un_operator)));
+                instr_stack.push(AgendaInstrs::Expr(un_operand));
             },
             PrimitiveOperation::BinaryOperation { operator, first_operand, second_operand } => {
-                first_operand.evaluate(instr_stack, stash);
-                second_operand.evaluate(instr_stack, stash);
-                operator.evaluate(instr_stack, stash);
+                let bin_operator = BinOp {
+                    sym: operator.clone()
+                };
+                let lhs_operand = first_operand.clone();
+                let rhs_operand = second_operand.clone();
+                instr_stack.push(AgendaInstrs::Instructions(Instructions::BinOp(bin_operator)));
+                instr_stack.push(AgendaInstrs::Expr(rhs_operand));
+                instr_stack.push(AgendaInstrs::Expr(lhs_operand));
             },
             PrimitiveOperation::VariadicOperation { operator, operands } => {
-                for operand in operands {
-                    operand.evaluate(instr_stack, stash);
-                }
-                operator.evaluate(instr_stack, stash);
+                // TODO
             },
-        }
-    }
-}
-
-
-impl Evaluate for UnaryOperator {
-    fn evaluate(&self, instr_stack: &mut Vec<AgendaInstrs>, stash: &mut Vec<Literal>) {
-        let operand = stash.pop().expect("Not enough operands on stack");
-        let result = match operand {
-            IntLiteral(x) => unop_microcode_num(x, *self),
-            BoolLiteral(value) => unop_microcode_bool(value, *self),
-            _ => panic!("UnaryOperator: Unsupported unary operations"),
-        };
-        stash.push(result);
-    }
-}
-
-impl Evaluate for BinaryOperator {
-    fn evaluate(&self, instr_stack: &mut Vec<AgendaInstrs>, stash: &mut Vec<Literal>) {
-        let rhs = stash.pop().expect("Not enough operands on stack");
-        let lhs = stash.pop().expect("Not enough operands on stack");
-        let result = match (lhs, rhs) {
-            (IntLiteral(x), IntLiteral(y)) => {
-                binop_microcode_num(x, y, *self)
-            }
-            (BoolLiteral(x), BoolLiteral(y)) => {
-                binop_microcode_bool(x, y, *self)
-            }
-            _ => panic!("BinaryOperator: Unsupported binary operation"),
-        };
-        stash.push(result);
-    }
-}
-
-impl Evaluate for VariadicOperator {
-    fn evaluate(&self, instr_stack: &mut Vec<AgendaInstrs>, stash: &mut Vec<Literal>) {
-        match self {
-            //todo!
-            _ => unimplemented!(),
         }
     }
 }
 
 impl Evaluate for Literal {
-    fn evaluate(&self, instr_stack: &mut Vec<AgendaInstrs>, stash: &mut Vec<Literal>) {
+    fn evaluate(&self, instr_stack: &mut Vec<AgendaInstrs>, stash: &mut Vec<Literal>, env: &mut Environment) {
         match self {
             IntLiteral(n) => stash.push(IntLiteral(*n)), // Need to  extract the literal when using it
             BoolLiteral(b) => stash.push(BoolLiteral(*b)),
             StringLiteral(s) => {}, // Heap
             UnitLiteral => { stash.push(UnitLiteral) },
+            _ => panic!("This literal type is not supported!")
         }
     }
 }
